Add useExpanderFocusClick hook
