+++ frontend/src/pages/RecipientRecord/pages/Monitoring/components/ReviewCard.js
@@ -1,4 +1,4 @@
-import React, { useState } from 'react';
+import React from 'react';
 import PropTypes from 'prop-types';
 import DataCard from '../../../../../components/DataCard';
 import DescriptionItem from './DescriptionItem';
@@ -7,21 +7,10 @@ import FindingWithinReview from './FindingWithinReview';
 import SpecialistTags from './SpecialistTags';
 import Tag from '../../../../../components/Tag';
 import ExpanderButton from '../../../../../components/ExpanderButton';
+import useExpanderFocusClick from '../../../../../hooks/useExpanderFocusClick';
 
 export default function ReviewCard({ review, regionId }) {
-  const [expanded, setExpanded] = useState(false);
-  const btnRef = React.useRef(null);
-
-  const handleExpanderClick = () => {
-    setExpanded(!expanded);
-
-    // Setting a timeout to ensure the button is focused after state update
-    setTimeout(() => {
-      if (btnRef.current) {
-        btnRef.current.focus();
-      }
-    }, 200);
-  };
+  const { expanded, btnRef, handleExpanderClick } = useExpanderFocusClick();
 
   return (
     <DataCard
